Minor revisions
Shorten generateLore() to genLore()
Ensure all references to variables use this if applicable
Move player cast outside the if else, this will mean it's casted anytime
an inventory click happens IF the slot type is the result and the
clicked item isnt air. So not often enough implication to consider
leaving it how it was.